morocco-economic-insights/requirements.txt
streamlit
plotly
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.16
langchain-groq==0.0.1
langchain-core>=0.1.23
pydantic>=2.7.0
pydantic-settings>=2.7.0
hnswlib==0.8.0
pandas==2.1.4
python-multipart==0.0.6
SQLAlchemy==2.0.25
sentence-transformers==2.5.1
numpy>=1.24.3
